updating code-tests to reflect the regrouping of checks on
a universal profile + vendor-specific profiles
(issue #2426)
